prediction module + container

+++ player-statistics/src/main/java/be/kdg/integration5/domain/Prediction.java
@@ -2,5 +2,5 @@ package be.kdg.integration5.domain;
 
 import java.time.LocalDate;
 
-public record Prediction(LocalDate date, int predictedMinutes) {
+public record Prediction(LocalDate date, double predictedMinutes) {
 }
